Рефакторинг: улучшение структуры классов и методов, добавление базового клиента, переименование методов для QR логина
