+++ README.md
@@ -14,9 +14,10 @@
 </p>
 
 > [!IMPORTANT]
-> (20.12.25) Из за резкого изменения апи большая часть библиотеки не работает. 
+> (20.12.25) Из за резкого изменения апи большая часть библиотеки не работает.
 Смотрите [новость](https://t.me/pymax_news/111)
-
+>
+> P.s добавил логин по qr в dev/1.2.1
 ---
 > ⚠️ **Дисклеймер**
 >
