+++ src/pymax/core.py
@@ -6,21 +6,19 @@ import logging
 import socket
 import ssl
 import time
-import traceback
 from collections.abc import Awaitable
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Literal
 from uuid import UUID
 
-from typing_extensions import Self, override
+from typing_extensions import override
 
 from .crud import Database
 from .exceptions import (
     InvalidPhoneError,
     SocketNotConnectedError,
-    WebSocketNotConnectedError,
 )
-from .formatter import ColoredFormatter
+from .interfaces import BaseClient
 from .mixins import ApiMixin, SocketMixin, WebSocketMixin
 from .payloads import UserAgentPayload
 from .static.constant import (
@@ -43,7 +41,7 @@ if TYPE_CHECKING:
 logger = logging.getLogger(__name__)
 
 
-class MaxClient(ApiMixin, WebSocketMixin):
+class MaxClient(ApiMixin, WebSocketMixin, BaseClient):
     """
     Основной клиент для работы с WebSocket API сервиса Max.
 
@@ -181,33 +179,11 @@ class MaxClient(ApiMixin, WebSocketMixin):
             self._work_dir,
         )
 
-    def _setup_logger(self) -> None:
-        if not self.logger.handlers:
-            if not self.logger.level:
-                self.logger.setLevel(logging.INFO)
-            handler = logging.StreamHandler()
-            formatter = ColoredFormatter(
-                "%(asctime)s [%(levelname)s] %(name)s: %(message)s",
-                datefmt="%Y-%m-%d %H:%M:%S",
-            )
-            handler.setFormatter(formatter)
-            self.logger.addHandler(handler)
-
     async def _wait_forever(self) -> None:
         try:
             await self.ws.wait_closed()
         except asyncio.CancelledError:
             self.logger.debug("wait_closed cancelled")
-
-    async def _safe_execute(self, coro, *, context: str = "unknown") -> Any:
-        """
-        Безопасно выполняет пользовательскую корутину.
-        Логирует traceback, но не роняет event loop.
-        """
-        try:
-            return await coro
-        except Exception as e:
-            self.logger.error(f"Unhandled exception in {context}: {e}\n{traceback.format_exc()}")
 
     async def close(self) -> None:
         """
@@ -236,24 +212,6 @@ class MaxClient(ApiMixin, WebSocketMixin):
         except Exception:
             self.logger.exception("Error closing client")
 
-    @override
-    def _create_safe_task(
-        self, coro: Awaitable[Any], *, name: str | None = None
-    ) -> asyncio.Task[Any | None]:
-        async def runner():
-            try:
-                return await coro
-            except asyncio.CancelledError:
-                raise
-            except Exception as e:
-                tb = traceback.format_exc()
-                self.logger.error(f"Unhandled exception in task {name or coro}: {e}\n{tb}")
-                raise
-
-        task = asyncio.create_task(runner(), name=name)
-        self._background_tasks.add(task)
-        return task
-
     async def _post_login_tasks(self, sync: bool = True) -> None:
         if sync:
             await self._sync()
@@ -276,44 +234,6 @@ class MaxClient(ApiMixin, WebSocketMixin):
             result = self._on_start_handler()
             if asyncio.iscoroutine(result):
                 await self._safe_execute(result, context="on_start handler")
-
-    async def _cleanup_client(self) -> None:
-        for task in list(self._background_tasks):
-            task.cancel()
-            try:
-                await task
-            except asyncio.CancelledError:
-                pass
-            except Exception:
-                self.logger.debug("Background task raised during cancellation", exc_info=True)
-            self._background_tasks.discard(task)
-
-        if self._recv_task:
-            self._recv_task.cancel()
-            with contextlib.suppress(asyncio.CancelledError):
-                await self._recv_task
-            self._recv_task = None
-
-        if self._outgoing_task:
-            self._outgoing_task.cancel()
-            with contextlib.suppress(asyncio.CancelledError):
-                await self._outgoing_task
-            self._outgoing_task = None
-
-        for fut in self._pending.values():
-            if not fut.done():
-                fut.set_exception(WebSocketNotConnectedError)
-        self._pending.clear()
-
-        if self._ws:
-            try:
-                await self._ws.close()
-            except Exception:
-                self.logger.debug("Error closing ws during cleanup", exc_info=True)
-            self._ws = None
-
-        self.is_connected = False
-        self.logger.info("Client start() cleaned up")
 
     async def login_with_code(self, temp_token: str, code: str, start: bool = False) -> None:
         """
@@ -397,43 +317,6 @@ class MaxClient(ApiMixin, WebSocketMixin):
             self.logger.info("Reconnect enabled — restarting client")
             await asyncio.sleep(self.reconnect_delay)
 
-    async def idle(self):
-        """
-        Поддерживает клиента в «ожидающем» состоянии до закрытия клиента или иного прерывающего события.
-
-        :return: Никогда не возвращает значение; функция блокирует выполнение.
-        :rtype: None
-        """
-        await asyncio.Event().wait()
-
-    def inspect(self) -> None:
-        """
-        Выводит в лог текущий статус клиента для отладки.
-        """
-        self.logger.info("Pymax")
-        self.logger.info("---------")
-        self.logger.info(f"Connected: {self.is_connected}")
-        if self.me is not None:
-            self.logger.info(f"Me: {self.me.names[0].first_name} ({self.me.id})")
-        else:
-            self.logger.info("Me: N/A")
-        self.logger.info(f"Dialogs: {len(self.dialogs)}")
-        self.logger.info(f"Chats: {len(self.chats)}")
-        self.logger.info(f"Channels: {len(self.channels)}")
-        self.logger.info(f"Users cached: {len(self._users)}")
-        self.logger.info(f"Background tasks: {len(self._background_tasks)}")
-        self.logger.info(f"Scheduled tasks: {len(self._scheduled_tasks)}")
-        self.logger.info("---------")
-
-    async def __aenter__(self) -> Self:
-        self._create_safe_task(self.start(), name="start")
-        while not self.is_connected:
-            await asyncio.sleep(0.05)
-        return self
-
-    async def __aexit__(self, exc_type, exc, tb) -> None:
-        await self.close()
-
 
 class SocketMaxClient(SocketMixin, MaxClient):
     @override
@@ -448,12 +331,6 @@ class SocketMaxClient(SocketMixin, MaxClient):
 
     @override
     async def _cleanup_client(self):
-        """
-        Socket-specific cleanup: cancel background tasks, set pending futures
-        exceptions to SocketNotConnectedError, and close socket.
-        """
-        from .exceptions import SocketNotConnectedError
-
         for task in list(self._background_tasks):
             task.cancel()
             try:
